+++ Changelog.md
@@ -1,3 +1,10 @@
+# 1.1.0
+
+* Add an upper bound on the amount of errors collected. This means that even
+  when things go horribly wrong in other code, you at least won't run out of
+  memory by collecting all the exceptions! Thanks to @gregorycollins for
+  this advice.
+
 # 1.0.0
 
 * Initial version.